requirements.txt
python-dotenv
chromadb
sentence-transformers
google-genai
requests
